UI & Level Manager Script
So, at this point it's just the pure basics on the canvas; simple title text and a couple of buttons that allow you to quit.

Also created the LevelManager gameobject, which is currently housing a couple of public functions that print stuff to the console when the buttons are tapped.

For the Start button, it's calling the LoadLevel function, which will eventually just select the right scene to switch to.

As for the Quit, it's calling the QuitRequest function, which will eventually just restart the game.