Button Logic
So, using the LevelManager script I've now got it hooked up so you can neatly jump between the different scenes.

Potentially thinking about branching out on the text-adventure game now to include this, as I'll then potentially be able to use a combo of states and scenes to make something really cool.

Note: Added scenes to the scene list in build settings - important as otherwise it doesn't recognise the scenes.