NumberWizard Imported
So, after ensuring the scenes could happily jump between each other and had the logic in place, I had to import in the Number Wizard script from the first tutorial.

A lot of it was now redundant, such as the printing to console, so I stripped it back and introduced a couple of public functions that enabled me to add them to the physical UI buttons within the scene.

Then also added in the ability to print the computer's guess to a string of UI text that's also in the scene.

That's pretty much the whole game done from a purely functional stand point. Next up will be to tidy it up into something a bit more presentable!